+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "cyezoi" extension will be documented in this file.
 
 ## v0.0.23
 
+- Enhance code render and Copy button
 - Fix Hydro Icons
 - Minify publish files